Update install_brew_fileuser.md
The goinfre folder is not in the home directory, it's in the /Volumes/Storage/ directory.
The school macs don't use bash, they use zsh, so using .bashrc_profile will not work. We simply have to add an alias to the binary file of brew inside our .zshrc file. It will be persistent.